+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## [0.1.34]
+* Add TenantId to app reg data
+* Rename UploadBlob to UploadBlobFromFile
+* Create UploadBlobFromString to allow handling of in memory data uploads
+* Update to Go 1.24.3
+* Update deps
+
 ## [0.1.33]
 * Bugfix: Add back type and subscription Name to IP address fetch
 * Bugfix: Cost data when some subscription names are caps has been resolved.